+++ plugin/evm/vm_test.go
@@ -3069,46 +3069,6 @@ func TestBuildInvalidBlockHead(t *testing.T) {
 	}
 }
 
-func TestConfigureLogLevel(t *testing.T) {
-	tests := []struct {
-		logLevel    string
-		expectedErr error
-	}{
-		{
-			logLevel: "info",
-		},
-		{
-			logLevel:    "cchain", // invalid
-			expectedErr: errInitializingLogger,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.logLevel, func(t *testing.T) {
-			require := require.New(t)
-
-			vm := &VM{}
-			ctx, dbManager, genesisBytes, issuer, _ := setupGenesis(t, upgradetest.Latest)
-			err := vm.Initialize(
-				context.Background(),
-				ctx,
-				dbManager,
-				genesisBytes,
-				nil,
-				fmt.Appendf(nil, `{"log-level": "%s"}`, test.logLevel),
-				issuer,
-				nil,
-				&enginetest.Sender{T: t},
-			)
-			require.ErrorIs(err, test.expectedErr)
-			if err != nil {
-				return
-			}
-
-			require.NoError(err, vm.Shutdown(context.Background()))
-		})
-	}
-}
-
 // Regression test to ensure we can build blocks if we are starting with the
 // Apricot Phase 4 ruleset in genesis.
 func TestBuildApricotPhase4Block(t *testing.T) {
